requirements.txt
gitdb>=4.0.7
GitPython>=3.1.18
smmap>=4.0.0
